requirements.txt

Django
djangorestframework
djangorestframework-authtoken
django-cors-headers
requests
torch
transformers
